Prevent recursive deletion via ~KdNode destructor
In kdTreeMapKnlogn.cpp, kdTreeKmapKnlog.cpp, kdTreeMapNlogn.cpp, and kdTreeKmapNlogn.cpp the ~KdNode destructor prevents recursive deletion of the duplicates list because that destructor deletes the duplicates list iteratively. In kdTreeKnlogn.cpp and kdTreeNlogn.cpp a comment is added to the removeDuplicates function to explain that any tuple not deleted by that function will eventually be deleted by the ~KdNode destructor.